LLM/Code/weather_api/requirements.txt
typer
requests
python-dotenv
rich